Added spread function and moved functions to its own file.
