+++ config.json
@@ -1,5 +1,5 @@
 {
 	"prefix": "!",
-    "token": "Nzk4Mzg3Njg5NzI4Mzc2ODQ1.X_0Seg.yGA-V1cx80noqIU7TKxaBjO_OHg"
+    "token": "Nzk4Mzg3Njg5NzI4Mzc2ODQ1.X_0Seg.wmEdrqXfyoH_9fQ4OmMoCaH-HgA"
 
 }
