+++ example-odds.json
@@ -10,7 +10,8 @@
           "site_key": "fanduel",
           "site_nice": "FanDuel",
           "last_update": "2020-10-07T09:13:27Z",
-          "odds": {"h2h": [188,-225]}
+          "odds": {"h2h": [188,-225],
+                   "spread":[3, -3] }
       },
       {
           "site_key": "bookmaker",
